Added two new features:
The ability to load complex data as a field in a model without it automatically treating it as hierarchical data if the model store doesn't know about a model with a matching name.

The ability to limit the scope of the turbolinkForms command to exclude parts of the page that should still go through usual page loads and redirects.